fix(types): move HDNodeResponse from protobuf to bitcoin file
