+++ src/js/core/methods/GetPublicKey.js
@@ -9,8 +9,7 @@ import { UiMessage } from '../../message/builder';
 
 import { getBitcoinNetwork } from '../../data/CoinInfo';
 import { getPublicKeyLabel } from '../../utils/pathUtils';
-import type { HDNodeResponse } from '../../types/trezor/protobuf';
-import type { CoreMessage, UiPromiseResponse, BitcoinNetworkInfo } from '../../types';
+import type { CoreMessage, UiPromiseResponse, BitcoinNetworkInfo, HDNodeResponse } from '../../types';
 
 type Batch = {
     path: Array<number>;
